test: add an automatic test runner
This commit removes the `spec` folder and adds an `auto_test_runner.lua`
file.

The `.busted` file is configured to only scan for this file and run it.

The auto test runner scans all the lua files, requires them and looks
for an exposed `__test` function. If it finds one, it calls it with the
global function environment (if we didn't override the fenv the `__test`
functions wouldn't have access to busted and luassert functions).

For modules which error upon `require` a special dummy test is made
which prints the error message, thus failing the entire suite.

This solution not only makes testing extremely flexible, it also allows
for easy testing of private implementation.
