+++ .busted
@@ -1,8 +1,10 @@
 return {
   _all = {
     coverage = false,
-    lpath = "lua/?.lua;lua/?/init.lua",
-    lua = "nlua",
+    lpath = 'lua/?.lua;lua/?/init.lua',
+		pattern = '^auto_test_runner.lua$',
+    lua = 'nlua',
+    ROOT = { 'lua/' }
   },
   default = {
     verbose = true
