+++ .luacheckrc
@@ -1,6 +1,6 @@
 std = "luajit"
 cache = true
-include_files = {"spec/**.lua", "lua/**.lua", "*.rockspec", ".luacheckrc"}
+include_files = {"lua/**.lua", "*.rockspec", ".luacheckrc"}
 
 read_globals = {
   "vim",
